+++ admin/core/Admin.php
@@ -158,10 +158,10 @@ class Admin extends Origin
 		$is_view = false;
 		if (support("table")) {
 			$is_view = is_view($tableStatus);
-			if ($is_view) {
-				$links["view"] = [lang('Alter view'), "edit"];
-			} else {
+			if (!$is_view) {
 				$links["create"] = [lang('Alter table'), "edit"];
+			} elseif (support("view")) {
+				$links["view"] = [lang('Alter view'), "edit"];
 			}
 		}